Desktop styles for header
