+++ components/common/Header/Header.module.scss
@@ -1,0 +1,25 @@
+@use './../../../styles/globals';
+
+.logo-style{
+    width: 48px;
+    height: 48px;
+    position: absolute;
+    top: 64px;
+    left: 55px;
+    img{
+        width: 100%;
+        height: 100%
+    }
+}
+
+.logo-style:after{
+    content: "";
+    height: 1px;
+    background-color: globals.$primary-color;
+    width: 473px;
+    opacity: 0.2514851987;
+    position: absolute;
+    top: 24px;
+    right: -537px;
+    z-index: 1;
+}
