+++ components/common/Layout/Layout.module.scss
@@ -1,0 +1,11 @@
+@use './../../../styles/globals';
+
+.app-container{
+    background-image: url('./../../../public/images/home/background-home-desktop.jpg');
+    background-repeat: no-repeat;
+    background-size:cover;
+    overflow: hidden;
+    width: 100vw;
+    height: 100vh;
+    min-height: 900px;
+}
