+++ components/common/Layout/layout.module.css
@@ -1,4 +1,4 @@
-.container {
+/* .container {
     max-width: 36rem;
     padding: 0 1rem;
     margin: 3rem auto 6rem;
@@ -12,4 +12,4 @@
   
   .backToHome {
     margin: 3rem 0 0;
-  }
+  } */
