+++ components/common/NavigationMenu/NavigationMenu.module.scss
@@ -8,3 +8,29 @@
 .heading{
     @include globals.main-heading(150px)
 }
+
+.navigation-menu{
+    width: 706px;
+    height: 56px;
+    position: absolute;
+    top: 40px;
+    right: 0px;
+    padding-left: 124px;
+    padding-top: 40px;
+    background: rgba(255, 255, 255, 0.04);
+    backdrop-filter: blur(40.774227142333984px);
+    @include globals.main-subheading(16px, 2.7px);
+    li{
+        display: inline-block;
+        vertical-align: top;
+        margin-right: 52px;
+        span{
+            font-weight: 700;
+            margin-right: 10px;
+        }
+        a{
+            color:inherit;
+            text-decoration: none;
+        }
+    }
+}
