+++ pages/index.js
@@ -1,10 +1,11 @@
 import Head from 'next/head';
-import Layout, { siteTitle } from '../components/layout';
-import utilStyles from '../styles/utils.module.css';
+import Layout, { siteTitle } from '../components/common/Layout/Layout';
+// import utilStyles from '../styles/utils.module.css';
 import { getSortedPostsData } from '../lib/posts';
 import Link from 'next/link';
 import Date from '../components/date';
 import { app_constants } from '../lib/data/globals';
+import styles from './../styles/Home.module.scss'
 
 export default function Home({ homeData }) {
